+++ src/main/java/com/clova/anifriends/domain/auth/LoginUser.java
@@ -1,4 +1,4 @@
-package com.clova.anifriends.domain.auth.resolver;
+package com.clova.anifriends.domain.auth;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
