+++ src/main/java/com/clova/anifriends/domain/auth/dto/response/LoginResponse.java
@@ -1,7 +1,6 @@
-package com.clova.anifriends.domain.auth.controller.response;
+package com.clova.anifriends.domain.auth.dto.response;
 
 import com.clova.anifriends.domain.auth.jwt.UserRole;
-import com.clova.anifriends.domain.auth.jwt.response.TokenResponse;
 
 public record LoginResponse(Long userId, UserRole role, String accessToken) {
