+++ src/main/java/com/clova/anifriends/domain/auth/dto/response/TokenResponse.java
@@ -1,4 +1,4 @@
-package com.clova.anifriends.domain.auth.jwt.response;
+package com.clova.anifriends.domain.auth.dto.response;
 
 import com.clova.anifriends.domain.auth.jwt.UserRole;
