+++ src/test/java/com/clova/anifriends/domain/auth/jwt/JJwtProviderTest.java
@@ -7,9 +7,10 @@ import static org.awaitility.Awaitility.await;
 import com.clova.anifriends.domain.auth.exception.ExpiredAccessTokenException;
 import com.clova.anifriends.domain.auth.exception.ExpiredRefreshTokenException;
 import com.clova.anifriends.domain.auth.exception.InvalidJwtException;
-import com.clova.anifriends.domain.auth.jwt.response.CustomClaims;
-import com.clova.anifriends.domain.auth.jwt.response.TokenResponse;
+import com.clova.anifriends.domain.auth.dto.response.CustomClaims;
+import com.clova.anifriends.domain.auth.dto.response.TokenResponse;
 import com.clova.anifriends.domain.auth.support.AuthFixture;
+import com.clova.anifriends.global.jwt.JJwtProvider;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 import org.awaitility.core.ThrowingRunnable;
