+++ src/test/java/com/clova/anifriends/domain/auth/resolver/LoginUserTestController.java
@@ -1,5 +1,6 @@
 package com.clova.anifriends.domain.auth.resolver;
 
+import com.clova.anifriends.domain.auth.LoginUser;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
