+++ src/test/java/com/clova/anifriends/domain/auth/support/AuthFixture.java
@@ -1,9 +1,9 @@
 package com.clova.anifriends.domain.auth.support;
 
-import com.clova.anifriends.domain.auth.jwt.JJwtProvider;
+import com.clova.anifriends.global.jwt.JJwtProvider;
 import com.clova.anifriends.domain.auth.jwt.JwtProvider;
 import com.clova.anifriends.domain.auth.jwt.UserRole;
-import com.clova.anifriends.domain.auth.jwt.response.TokenResponse;
+import com.clova.anifriends.domain.auth.dto.response.TokenResponse;
 
 public final class AuthFixture {
